refactor: 중복 엔티티 제거 및 장바구니 와 아이템 옵션간의 관계 삭제
