--- src/main/java/techit/gongsimchae/domain/category/repository/CategoryRepository.java
@@ -1,10 +1,0 @@
-package techit.gongsimchae.domain.category.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import techit.gongsimchae.domain.category.entity.Category;
-
-import java.util.Optional;
-
-public interface CategoryRepository extends JpaRepository<Category, Long> {
-    Optional<Category> findByName(String categoryName);
-}
