--- src/main/java/techit/gongsimchae/domain/groupbuying/category/dto/CategoryDto.java
@@ -1,4 +1,0 @@
-package techit.gongsimchae.domain.groupbuying.category.dto;
-
-public class CategoryDto {
-}
