+++ src/main/java/techit/gongsimchae/domain/groupbuying/category/repository/CategoryRepository.java
@@ -2,6 +2,10 @@ package techit.gongsimchae.domain.groupbuying.category.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import techit.gongsimchae.domain.groupbuying.category.entity.Category;
+import techit.gongsimchae.domain.groupbuying.category.entity.Category;
 
-public interface CategoryRepository extends JpaRepository<Category,Long> {
+import java.util.Optional;
+
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    Optional<Category> findByName(String categoryName);
 }
