--- src/main/java/techit/gongsimchae/domain/groupbuying/coupon/dto/CouponDto.java
@@ -1,4 +1,0 @@
-package techit.gongsimchae.domain.groupbuying.coupon.dto;
-
-public class CouponDto {
-}
