--- src/main/java/techit/gongsimchae/domain/groupbuying/coupon/service/CouponService.java
@@ -1,4 +1,0 @@
-package techit.gongsimchae.domain.groupbuying.coupon.service;
-
-public class CouponService {
-}
