--- src/main/java/techit/gongsimchae/domain/groupbuying/post/entity/PostEntity.java
@@ -1,4 +1,0 @@
-package techit.gongsimchae.domain.groupbuying.post.entity;
-
-public class PostEntity {
-}
